feat(web): Implement initial view mode, with canvas and left sidebar
Extras:
- Removed dependency on PixiJS for decoding component colors
- Refactored some WorkspaceEditor rxjs to be more idiomatic
